requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.4
giphy-client==1.0.0
idna==2.10
multidict==4.7.6
pprint==0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
tabulate==0.8.7
tmdbsimple==2.5.5
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
